requirements.txt
absl-py==0.7.1
astor==0.8.0
backcall==0.1.0
chardet==2.3.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.4
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
imageio==2.5.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
kiwisolver==1.1.0
lmdb
lxml==4.3.3
Markdown==3.1.1
matplotlib==2.2.4
mock==3.0.5
networkx==2.3
numpy==1.17.4
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
protobuf==3.11.1
ptyprocess==0.6.0
pyarrow==0.15.1
pycurl==7.43.0
Pygments==2.4.2
pygobject==3.20.0
pyparsing==2.4.0
python-dateutil==2.8.1
python-debian
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.0
six==1.13.0
tensorboard==1.13.1
tensorboardX==1.9
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
